Fixes double noise algorithms due to the EvaluationJobs by removing the algorithm from the noise class and creating a Evaluation job even when evaluating a single Value.

+++ Masterthesis/Assets/Scripts/Noise/LayeredNoise/NoiseLayerSettings.cs
@@ -10,8 +10,8 @@ namespace SBaier.Master
 		public NoiseSettings NoiseSettings => _noiseSettings;
 
 		[SerializeField]
-		private double _weight = 1;
-		public double Weight => _weight;
+		private float _weight = 1;
+		public float Weight => _weight;
 
 		[SerializeField]
 		private float _frequencyFactor = 1;
